+++ plugins/test/ipaddress_tests.py
@@ -14,13 +14,6 @@ from ansible_collections.ansible.utils.plugins.module_utils.common.ipaddress_uti
 
 __metaclass__ = type
 
-DOCUMENTATION = """
-    name: general
-    author: Bradley A. Thornton (@cidrblock)
-    version_added: "2.0.1"
-"""
-
-
 # import json
 import re
 
@@ -98,14 +91,14 @@ import re
 #         raise AnsibleError(msg)
 
 
-@_need_ipaddress
-def _in_network(ip, network):
-    """Test if an address or network is in a network<br/>`'10.1.1.1' is ansible.utils.in_network '10.0.0.0/8'`
-    """
-    try:
-        return _is_subnet_of(ip_network(ip), ip_network(network))
-    except Exception:
-        return False
+# @_need_ipaddress
+# def _in_network(ip, network):
+#     """Test if an address or network is in a network<br/>`'10.1.1.1' is ansible.utils.in_network '10.0.0.0/8'`
+#     """
+#     try:
+#         return _is_subnet_of(ip_network(ip), ip_network(network))
+#     except Exception:
+#         return False
 
 
 def _in_one_network(ip, networks):
